fix: allow composite projects
We need to propagate the tsconfig.json file to composite project dependents as an input.
This was in rules_nodejs but we lost it.

Also correct the pathing for the tsbuildinfo output, which hadn't been tested in rules_ts yet.

Copied the three composite projects from rules_nodejs tests, so we have that coverage now.

Fixes #163