fix: ts_project external workspace builds (#310)
* fix: remove bin dir prefix instead of using `short_path`

Refs #284.

This allows `ts_project()` to build correctly while in an external workspace. This commit is a slightly modified version of [this diff](https://github.com/aspect-build/rules_ts/issues/284#issuecomment-1386030357).

* test: add a regression test using `ts_project()` in an external workspace

Closes #284.

This sets up two workspaces, one with a `ts_project()` and another which depends on the first. When the second workspace is built, it triggers a build of the `ts_project()` which is under `@lib_wksp`, not the main workspace. This would trigger #284 if not properly fixed. Note that we don't actually need to do anything with the `ts_project()` output, we just need to trigger to build, so a `build_test()` is sufficient to reproduce the error.