fix: wire the validation action
Sadly in #138 we only did this halfway: it stopped depending on the output of the validation action, but didn't actually execute one either. It has been 'failed open' this whole time, so likely users have ts_project attributes that have drifted from the tsconfig settings.

BREAKING CHANGE:
the validator is enabled again, so it will complain about ts_project attributes that don't match tsconfig.
You should just run the buildozer commands it prints to repair this.