chore: enable lib checks
The 'skipLibCheck=always' was added in https://github.com/aspect-build/rules_ts/pull/371 without any explanation of why we'd want that setting for rules_ts itself.
This is a prefactor for the regression test of #516 - and in general we should be testing with the 'stricter' setting to make more assurance that the rules behave correctly.