chore: tear out check_for_updates feature
It hasn't resulted in anyone knowing that a rules_ts upgrade was available.
The CDN traffic from this is the vast majority of what we are serving.
We never decided to roll it out to any other rulesets, so I feel that the experiment has failed.
I also never finished the server-side code that responds to the requests.