fix: do not include `tsconfig.json` files in `JsInfo` types
It's unclear to me what the `tsconfig.json` files from `ts_config`
targets end up contributing to the JSInfo type declarations. Those
files shouldn't be necessary for represent the output of a TS
compilation in general.

I assume, they could be necessary for composite projects, but even then
I'd expect the tsconfig to be retrieved in consuming targets via the
dedicated `TsConfigInfo` provider.