feat: support custom progress messages (#807)
For Angular we fully rely on `ts_project` just with a custom `tsc`
executable. We would like to have better distinction between Angular or
TypeScript targets.

It seems like it would be useful to control the progress message. This
would also allow better distinction when e.g. compiling with a different
configuration. e.g. we use a transition for Angular to compile with a
different compilation mode when shipping to npm