fix: update build_test() fork for compatibility with Bazel 8.3.0
We here imitate the changes in bazel_skylib 1.8.0 to avoid referring
to `@bazel_tools//src/conditions:host_windows`.  As of Bazel 8.3.0
referring to `@bazel_tools//src/conditions:host_windows` yields a
deprecation warning.

In concert with this, we here bump the minimum required version
of bazel_skylib to 1.8.0.  Per the comments in ts/repositories.bzl
this may be a BREAKING CHANGE.

Fixes #824