fix: allow out_dir to be an empty string in validator
ts_project sets the out_dir to None if it's "." or "./" which makes it
as if it is not set or effectively an empty string. The validator then
checks of the ts_project rule has an out_dir set if the tsconfig has it
set, but it got stripped by the ts_project thus the validator fails with
an error saying the tsconfig doesn't match the rule - when it actually
does.